+++ courses/coding-in-C++/Lab_4/my_solution/section1/vector2d.hpp
@@ -13,7 +13,7 @@ public:
     void print();
 
     double magnitude();
-    int magnitude(int precision);
+    double magnitude(int precision);
 };
 
 #endif
